Integración de backend con Stripe y configuración para despliegue en Vercel
