+++ .gitignore
@@ -1,4 +1,4 @@
-# Logs
+z# Logs
 logs
 *.log
 npm-debug.log*
