+++ vercel.json
@@ -2,14 +2,22 @@
   "version": 2,
   "builds": [
     {
-      "src": "api/server.js",
+      "src": "backend/api/server.js",
       "use": "@vercel/node"
+    },
+    {
+      "src": "dist/**",
+      "use": "@vercel/static"
     }
   ],
   "routes": [
     {
+      "src": "/api/(.*)",
+      "dest": "backend/api/server.js"
+    },
+    {
       "src": "/(.*)",
-      "dest": "api/server.js"
+      "dest": "/index.html"
     }
   ]
 }
